+++ redfish_interop_validator/interop.py
@@ -3,26 +3,31 @@
 # Copyright 2016 DMTF. All rights reserved.
 # License: BSD 3-Clause License. For full text see link: https://github.com/DMTF/Redfish-Interop-Validator/blob/master/LICENSE.md
 
-import re, copy
+import re
 from enum import Enum
 from collections import Counter
 
 import logging
-from common.redfish import getNamespaceUnversioned, getType, getNamespace
-from traverseInterop import callResourceURI
+from redfish_interop_validator.redfish import getNamespaceUnversioned, getType, getNamespace
+from redfish_interop_validator.traverseInterop import callResourceURI
 my_logger = logging.getLogger()
 my_logger.setLevel(logging.DEBUG)
 
 config = {'WarnRecommended': False, 'WriteCheck': False}
 
-class sEnum(Enum):
+
+class testResultEnum(Enum):
     FAIL = 'FAIL'
     NOPASS = 'NO PASS'
     PASS = 'PASS'
     WARN = 'WARN'
     OK = 'OK'
+    NA = 'N/A'
+    NOT_TESTED = 'NOT TESTED'
+
 
 REDFISH_ABSENT = 'n/a'
+
 
 class msgInterop:
     def __init__(self, name, profile_entry, expected, actual, success):
@@ -32,7 +37,7 @@ class msgInterop:
         self.actual = actual
         self.ignore = False
         if isinstance(success, bool):
-            self.success = sEnum.PASS if success else sEnum.FAIL
+            self.success = testResultEnum.PASS if success else testResultEnum.FAIL
         else:
             self.success = success
         self.parent_results = None
@@ -66,33 +71,33 @@ def validateComparisonAnyOfAllOf(profile_entry, property_path="Unspecified"):
             # OK if passing, FAIL if check fails and value is not in array
             for msg in my_msgs:
                 msg.ignore = False
-                msg.success = sEnum.NOPASS
+                msg.success = testResultEnum.NOPASS
                 msg.expected = '{} {} ({})'.format(msg.expected, expected_values, "Across All Resources")
 
             if my_compare == 'AnyOf':
                 if any([x in my_values for x in expected_values]):
                     my_logger.info('  PASS')
-                    top_msg.success = sEnum.PASS
+                    top_msg.success = testResultEnum.PASS
                     for msg in my_msgs:
-                        msg.success = sEnum.OK
+                        msg.success = testResultEnum.OK
                         if msg.actual in expected_values:
-                            msg.success = sEnum.PASS
+                            msg.success = testResultEnum.PASS
                 else:
                     my_logger.info('  FAIL')
                     for msg in my_msgs:
-                        msg.success = sEnum.FAIL
+                        msg.success = testResultEnum.FAIL
 
             if my_compare == 'AllOf':
                 if all([x in my_values for x in expected_values]):
                     my_logger.info('  PASS')
-                    top_msg.success = sEnum.PASS
+                    top_msg.success = testResultEnum.PASS
                     for msg in my_msgs:
-                        msg.success = sEnum.OK
+                        msg.success = testResultEnum.OK
                 else:
                     my_logger.info('  FAIL')
                     for msg in my_msgs:
                         if msg.actual not in expected_values:
-                            msg.success = sEnum.FAIL
+                            msg.success = testResultEnum.FAIL
 
         if property_profile.get('PropertyRequirements'):
             new_msgs = validateComparisonAnyOfAllOf(property_profile.get('PropertyRequirements'), '.'.join([property_path, key]))
@@ -112,6 +117,7 @@ def validateRequirement(profile_entry, rf_payload_item=None, conditional=False, 
     # If we're not mandatory, pass automatically, else fail
     # However, we have other entries "IfImplemented" and "Conditional"
     # note: Mandatory is default!! if present in the profile.  Make sure this is made sure.
+    # For DNE entries "IfImplemented" and "Recommended" result with not applicable
     original_profile_entry = profile_entry
 
     if profile_entry == "IfPopulated":
@@ -126,15 +132,23 @@ def validateRequirement(profile_entry, rf_payload_item=None, conditional=False, 
 
     if profile_entry == "Conditional" and conditional:
         profile_entry = "Mandatory"
-    if profile_entry == "IfImplemented":
-        my_logger.debug('\tItem cannot be tested for Implementation')
+
     paramPass = not profile_entry == "Mandatory" or \
         profile_entry == "Mandatory" and not propDoesNotExist
+
+    if profile_entry == "IfImplemented":
+        if propDoesNotExist:
+            paramPass = testResultEnum.NA
+        else:
+            my_logger.debug('\tItem cannot be tested for Implementation')
+
     if profile_entry == "Recommended" and propDoesNotExist:
         my_logger.info('\tItem is recommended but does not exist')
         if config['WarnRecommended']:
             my_logger.warning('\tItem is recommended but does not exist, escalating to WARN')
-            paramPass = sEnum.WARN
+            paramPass = testResultEnum.WARN
+        else:
+            paramPass = testResultEnum.NA
 
     my_logger.debug('\tpass ' + str(paramPass))
     return msgInterop('ReadRequirement', original_profile_entry, 'Must Exist' if profile_entry == "Mandatory" else 'Any', 'Exists' if not propDoesNotExist else 'DNE', paramPass),\
@@ -186,34 +200,52 @@ def findPropItemforString(propObj, itemname):
     return None
 
 
-def validateWriteRequirement(propObj, profile_entry, itemname):
+def validateWriteRequirement(profile_entry, parent_object_payload, resource_headers, item_name):
     """
     Validates if a property is WriteRequirement or not
     """
-    my_logger.debug('writeable \n\t' + str(profile_entry))
-    permission = 'Read'
-    expected = "OData.Permission/ReadWrite" if profile_entry else "Any"
-    if not config['WriteCheck']:
-        paramPass = True
-        return msgInterop('WriteRequirement', profile_entry, expected, permission, paramPass),\
-            paramPass
-    if profile_entry:
-        targetProp = findPropItemforString(propObj, itemname.replace('#', ''))
-        propAttr = None
-        if targetProp is not None:
-            propAttr = targetProp.propDict.get('OData.Permissions')
-        if propAttr is not None:
-            permission = propAttr.get('EnumMember', 'Read')
-            paramPass = permission \
-                == "OData.Permission/ReadWrite"
-        else:
-            paramPass = False
-    else:
-        paramPass = True
+    my_logger.verbose1('Is property writeable \n\t' + str(profile_entry))
 
-    my_logger.debug('\tpass ' + str(paramPass))
-    return msgInterop('WriteRequirement', profile_entry, expected, permission, paramPass),\
-        paramPass
+    if profile_entry == "Mandatory" or profile_entry == "Supported":
+        result_not_supported = testResultEnum.FAIL
+        expected_str = "Must Be Writable"
+    elif profile_entry == "Recommended":
+        if config['WarnRecommended']:
+            result_not_supported = testResultEnum.WARN
+        else:
+            result_not_supported = testResultEnum.NA
+        expected_str = "Recommended"
+    else:
+        result_not_supported = testResultEnum.NA
+        expected_str = "Any"
+
+    # Check for Allow header, warn if missing
+    if resource_headers and 'Allow' in resource_headers:
+        writeable = 'PATCH' in resource_headers['Allow']
+        if not writeable:
+            if profile_entry == "Mandatory":
+                my_logger.error('PATCH in Allow header not available, property is not writeable ' + str(profile_entry))
+            return msgInterop('WriteRequirement', profile_entry, expected_str, 'PATCH not supported', result_not_supported), True
+    else:
+        my_logger.warning('Unable to test writeable property, no Allow header available ' + str(profile_entry))
+        return msgInterop('WriteRequirement', profile_entry, expected_str, 'No Allow response header', testResultEnum.NOT_TESTED), True
+    
+    redfish_payload, _ = parent_object_payload
+
+    # Get Writeable Properties
+    if '@Redfish.WriteableProperties' not in redfish_payload:
+        my_logger.warning('Unable to test writeable property, no @Redfish.WriteableProperties available at the property level ' + str(profile_entry))
+        return msgInterop('WriteRequirement', profile_entry, expected_str, '@Redfish.WriteableProperties not in response', testResultEnum.NOT_TESTED), True
+
+    writeable_properties = redfish_payload['@Redfish.WriteableProperties']
+    if not isinstance(writeable_properties, list):
+        my_logger.warning('Unable to test writeable property, @Redfish.WriteableProperties is not an array ' + str(profile_entry))
+        return msgInterop('WriteRequirement', profile_entry, expected_str, '@Redfish.WriteableProperties not an array', testResultEnum.WARN), True
+
+    is_writeable = item_name in writeable_properties
+
+    return msgInterop('WriteRequirement', profile_entry, expected_str, 'Writable' if is_writeable else 'Not Writable',
+                      testResultEnum.PASS if is_writeable else result_not_supported), True
 
 
 def checkComparison(val, compareType, target):
@@ -247,25 +279,29 @@ def checkComparison(val, compareType, target):
             else:
                 continue
         paramPass = len(alltarget) == len(target)
+
     if compareType == "LinkToResource":
-        vallink = val.get('@odata.id')
-        success, rf_payload, code, elapsed = callResourceURI(vallink)
-        if success:
-            ourType = rf_payload.get('@odata.type')
-            if ourType is not None:
-                SchemaType = getType(ourType)
-                paramPass = SchemaType in target
+        if val == REDFISH_ABSENT:
+            paramPass = False
+        else:
+            vallink = val.get('@odata.id')
+            success, rf_payload, code, elapsed, _ = callResourceURI(vallink)
+            if success:
+                ourType = rf_payload.get('@odata.type')
+                if ourType is not None:
+                    SchemaType = getType(ourType)
+                    paramPass = SchemaType in target
+                else:
+                    paramPass = False
             else:
                 paramPass = False
-        else:
-            paramPass = False
 
     if compareType == "Absent":
         paramPass = val == REDFISH_ABSENT
     if compareType == "Present":
         paramPass = val != REDFISH_ABSENT
 
-    if isinstance(target, list):
+    if isinstance(target, list) and val != REDFISH_ABSENT:
         if compareType == "Equal":
             paramPass = val in target
         elif compareType == "NotEqual":
@@ -282,6 +318,10 @@ def checkComparison(val, compareType, target):
                     paramPass = val <= value
                 if paramPass is False:
                     break
+    elif compareType in ["Equal", "NotEqual", "GreaterThan", "GreaterThanOrEqual", "LessThan", "LessThanOrEqual"]:
+        if not isinstance(target, list):
+            my_logger.warn('CompareType {} requires a list of values'.format(compareType))
+
     my_logger.debug('\tpass ' + str(paramPass))
     return msgInterop('Comparison', target, compareType, val, paramPass),\
         paramPass
@@ -399,6 +439,7 @@ def checkConditionalRequirement(propResourceObj, profile_entry, rf_payload_tuple
         my_logger.error("Invalid Profile - No conditional given")
         raise ValueError('No conditional given for Comparison')
 
+
 def find_key_in_payload(path_to_key, redfish_parent_payload):
     """
     Finds a key in the payload tuple provided
@@ -435,7 +476,8 @@ def find_key_in_payload(path_to_key, redfish_parent_payload):
             key_exists = False
     return key_exists
 
-def validatePropertyRequirement(propResourceObj, profile_entry, rf_payload_tuple, item_name, chkCondition=False):
+
+def validatePropertyRequirement(propResourceObj, profile_entry, rf_payload_tuple, item_name):
     """
     Validate PropertyRequirements
     """
@@ -455,7 +497,7 @@ def validatePropertyRequirement(propResourceObj, profile_entry, rf_payload_tuple
         replacement_property_exists = find_key_in_payload(my_path_entry, redfish_parent_payload)
 
         new_msg = msgInterop("{}.{}".format(item_name, "ReplacesProperty"), profile_entry["ReplacesProperty"], "-",
-                            "Exists" if replacement_property_exists else "DNE", sEnum.WARN if replacement_property_exists else sEnum.OK)
+                            "Exists" if replacement_property_exists else "DNE", testResultEnum.WARN if replacement_property_exists else testResultEnum.OK)
         msgs.append(new_msg)
         if replacement_property_exists:
             my_logger.warn('{}: This property replaces deprecated property {}, but does not exist, service should implement {}'.format(item_name, my_path_entry, item_name))
@@ -469,7 +511,7 @@ def validatePropertyRequirement(propResourceObj, profile_entry, rf_payload_tuple
         replacement_property_exists = find_key_in_payload(my_path_entry, redfish_parent_payload)
         
         new_msg = msgInterop("{}.{}".format(item_name, "ReplacedByProperty"), profile_entry["ReplacedByProperty"], "-",
-                            "Exists" if replacement_property_exists else "DNE", sEnum.PASS if replacement_property_exists else sEnum.OK)
+                            "Exists" if replacement_property_exists else "DNE", testResultEnum.PASS if replacement_property_exists else testResultEnum.OK)
         msgs.append(new_msg)
         if replacement_property_exists:
             my_logger.info('{}: Replacement property exists, step out of validating'.format(item_name))
@@ -486,7 +528,7 @@ def validatePropertyRequirement(propResourceObj, profile_entry, rf_payload_tuple
                 if checkConditionalRequirement(propResourceObj, item, rf_payload_tuple):
                     my_logger.info("\tCondition DOES apply")
                     conditionalMsgs, conditionalCounts = validatePropertyRequirement(
-                        propResourceObj, item, rf_payload_tuple, item_name, chkCondition=True)
+                        propResourceObj, item, rf_payload_tuple, item_name)
                     counts.update(conditionalCounts)
                     for item in conditionalMsgs:
                         item.name = item.name.replace('.', '.Conditional.', 1)
@@ -521,12 +563,14 @@ def validatePropertyRequirement(propResourceObj, profile_entry, rf_payload_tuple
         # problem: if dne, skip?
 
         # Read Requirement is default mandatory if not present
-        msg, success = validateRequirement(profile_entry.get('ReadRequirement', 'Mandatory'), redfish_value, parent_object_tuple=redfish_parent_payload)
+        requirement_entry = profile_entry.get('ReadRequirement', 'Mandatory')
+        msg, requirement_success = validateRequirement(requirement_entry, redfish_value, parent_object_tuple=redfish_parent_payload)
         msgs.append(msg)
         msg.name = item_name + '.' + msg.name
 
         if "WriteRequirement" in profile_entry:
-            msg, success = validateWriteRequirement(propResourceObj, profile_entry["WriteRequirement"], item_name)
+            headers = propResourceObj.headers
+            msg, success = validateWriteRequirement(profile_entry.get('WriteRequirement', 'Mandatory'), redfish_parent_payload, headers, item_name)
             msgs.append(msg)
             msg.name = item_name + '.' + msg.name
             if not success:
@@ -541,12 +585,16 @@ def validatePropertyRequirement(propResourceObj, profile_entry, rf_payload_tuple
             if not success:
                 my_logger.error("MinSupportValues failed")
 
-        if "Values" in profile_entry and not chkCondition:
+        if "Values" in profile_entry:
             # Default to AnyOf
-            # NOTE: chkCondition seems to skip this if a ConditionalRequirement is met, this may be unnecessary
 
             my_compare = profile_entry.get("Comparison", "AnyOf")
-            msg, success = checkComparison(redfish_value, my_compare, profile_entry.get("Values", []))
+            my_values = profile_entry.get("Values", [])
+            # If absent and not comparing for absence...
+            if redfish_value == REDFISH_ABSENT and my_compare not in ['Absent', 'Present']:
+                msg, success = msgInterop('Comparison', my_values, my_compare, REDFISH_ABSENT, testResultEnum.NOT_TESTED), True
+            else:
+                msg, success = checkComparison(redfish_value, my_compare, my_values)
             msgs.append(msg)
             msg.name = item_name + '.' + msg.name
 
@@ -584,72 +632,119 @@ def validateActionRequirement(profile_entry, rf_payload_tuple, actionname):
     my_logger.verbose1('actionRequirement \n\tval: ' + str(rf_payload_item if not isinstance(
         rf_payload_item, dict) else 'dict') + ' ' + str(profile_entry))
 
+    action_readrequirement = profile_entry.get('ReadRequirement', "Mandatory")
+    actioninfo_requirement = profile_entry.get('ActionInfo', "None")
+
     if "ReadRequirement" in profile_entry:
         # problem: if dne, skip
-        msg, success = validateRequirement(profile_entry.get('ReadRequirement', "Mandatory"), rf_payload_item)
+        msg, success = validateRequirement(action_readrequirement, rf_payload_item)
         msgs.append(msg)
         msg.name = actionname + '.' + msg.name
-        msg.success = sEnum.PASS if success else sEnum.FAIL
-
+        msg.success = testResultEnum.PASS if success else testResultEnum.FAIL
+    
     propDoesNotExist = (rf_payload_item == REDFISH_ABSENT)
     if propDoesNotExist:
         return msgs, counts
+
     if "@Redfish.ActionInfo" in rf_payload_item:
         vallink = rf_payload_item['@Redfish.ActionInfo']
-        success, rf_payload_action, code, elapsed = callResourceURI(vallink)
+        success, rf_payload_action, code, elapsed, _ = callResourceURI(vallink)
         if not success:
             rf_payload_action = None
 
+    if 'ActionInfo' in profile_entry and actioninfo_requirement in ["None"]:
+        # Create message if None is explicitly listed in the profile
+        msg = msgInterop('ActionInfo', 'None', '-', '-', testResultEnum.OK)
+        msg.name = actionname + '.' + msg.name
+        msgs.append(msg)
+
+    if actioninfo_requirement not in ["None"]:
+        if propDoesNotExist:
+            # not tested if action isn't present
+            msg = msgInterop('ActionInfo', actioninfo_requirement, '-', '-', testResultEnum.NOT_TESTED)
+
+        elif actioninfo_requirement == "Mandatory":
+            if rf_payload_action is None:
+                if "@Redfish.ActionInfo" in rf_payload_item:
+                    my_logger.error('Mandatory @Redfish.ActionInfo for {} listed on action but URI get was not successful'.format(actionname))
+                else:
+                    my_logger.error('@Redfish.ActionInfo for {} not listed, but is Mandatory'.format(actionname))
+                msg = msgInterop('ActionInfo', actioninfo_requirement, '-', '-', testResultEnum.FAIL)
+            else:
+                msg = msgInterop('ActionInfo', actioninfo_requirement, '-', '-', testResultEnum.PASS)
+
+        elif actioninfo_requirement == "Recommended":
+            if rf_payload_action is None:
+                if "@Redfish.ActionInfo" in rf_payload_item:
+                    my_logger.warn('Recommended @Redfish.ActionInfo for {} listed on action but URI get was not successful'.format(actionname))
+                    msg = msgInterop('ActionInfo', actioninfo_requirement, '-', '-', testResultEnum.WARN)
+                else:
+                    my_logger.info('Recommended @Redfish.ActionInfo for {} not listed'.format(actionname))
+                    msg = msgInterop('ActionInfo', actioninfo_requirement, '-', '-', testResultEnum.PASS)
+            else:
+                msg = msgInterop('ActionInfo', actioninfo_requirement, '-', '-', testResultEnum.PASS)
+
+        else:
+            my_logger.warn('Term "ActionInfo" has unknown value {}'.format(actioninfo_requirement))
+            msg = msgInterop('ActionInfo', actioninfo_requirement, '-', '-', testResultEnum.WARN)
+        msg.name = actionname + '.' + msg.name
+        msgs.append(msg)
+
     # problem: if dne, skip
     if "Parameters" in profile_entry:
-        innerDict = profile_entry["Parameters"]
+        parameter_dictionary = profile_entry["Parameters"]
         # problem: if dne, skip
         # assume mandatory
-        for k in innerDict:
-            item = innerDict[k]
+        for param in parameter_dictionary:
+            item = parameter_dictionary[param]
+            # Get Allowable Values for parameter
             values_array = None
+            # If our action info exists at all, prefer it
             if rf_payload_action is not None:
-                action_by_name = rf_payload_action['Parameters']
-                my_action = [x for x in action_by_name if x['Name'] == k]
-                if my_action:
-                    values_array = my_action[0].get('AllowableValues')
+                parameter_by_name = rf_payload_action['Parameters']
+                my_parameter = [x for x in parameter_by_name if x['Name'] == param]
+                if my_parameter:
+                    values_array = my_parameter[0].get('AllowableValues')
+            # Otherwise check for AllowableValues as additional property
             if values_array is None:
-                values_array = rf_payload_item.get(str(k) + '@Redfish.AllowableValues', REDFISH_ABSENT)
+                values_array = rf_payload_item.get(str(param) + '@Redfish.AllowableValues', REDFISH_ABSENT)
             if values_array == REDFISH_ABSENT:
-                my_logger.warning('\tNo such ActionInfo exists for this Action, and no AllowableValues exists.  Cannot validate the following parameters: {}'.format(k))
-                msg = msgInterop('', item, '-', '-', sEnum.WARN)
-                msg.name = "{}.{}.{}".format(actionname, k, msg.name)
+                my_logger.warning('\tNo such ActionInfo exists for this Action, and no AllowableValues exists.  Cannot validate the following parameters: {}'.format(param))
+                msg = msgInterop('', item, '-', '-', testResultEnum.WARN)
+                msg.name = "{}.{}.{}".format(actionname, param, msg.name)
                 msgs.append(msg)
             else:
                 msg, success = validateRequirement(item.get('ReadRequirement', "Mandatory"), values_array)
                 msgs.append(msg)
-                msg.name = "{}.{}.{}".format(actionname, k, msg.name)
+                msg.name = "{}.{}.{}".format(actionname, param, msg.name)
                 if "ParameterValues" in item:
                     msg, success = validateSupportedValues(
                             item["ParameterValues"], values_array)
                     msgs.append(msg)
-                    msg.name = "{}.{}.{}".format(actionname, k, msg.name)
+                    msg.name = "{}.{}.{}".format(actionname, param, msg.name)
                 if "RecommendedValues" in item:
                     msg, success = validateSupportedValues(
                             item["RecommendedValues"], values_array)
                     msg.name = msg.name.replace('Supported', 'Recommended')
                     if config['WarnRecommended'] and not success:
                         my_logger.warning('\tRecommended parameters do not all exist, escalating to WARN')
-                        msg.success = sEnum.WARN
+                        msg.success = testResultEnum.WARN
                     elif not success:
                         my_logger.warning('\tRecommended parameters do not all exist, but are not Mandatory')
-                        msg.success = sEnum.PASS
+                        msg.success = testResultEnum.PASS
 
                     msgs.append(msg)
-                    msg.name = "{}.{}.{}".format(actionname, k, msg.name)
+                    msg.name = "{}.{}.{}".format(actionname, param, msg.name)
     # consider requirement before anything else, what if action
     # if the action doesn't exist, you can't check parameters
     # if it doesn't exist, what should not be checked for action
     return msgs, counts
 
+
 URI_ID_REGEX = '\{[A-Za-z0-9]*Id\}'
 
 VALID_ID_REGEX = '([A-Za-z0-9.!#$&-;=?\[\]_~])+'
+
 
 def compareRedfishURI(expected_uris, uri):
     success = False
@@ -662,6 +757,7 @@ def compareRedfishURI(expected_uris, uri):
         success = True
     return success
 
+
 def checkInteropURI(r_obj, profile_entry):
     """
     Checks if the profile's URI applies to the particular resource
@@ -670,6 +766,7 @@ def checkInteropURI(r_obj, profile_entry):
 
     my_id, my_uri = r_obj.jsondata.get('Id'), r_obj.uri
     return compareRedfishURI(profile_entry, my_uri)
+
 
 def validateInteropResource(propResourceObj, interop_profile, rf_payload):
     """
@@ -684,17 +781,27 @@ def validateInteropResource(propResourceObj, interop_profile, rf_payload):
 
     if "UseCases" in interop_profile:
         for use_case in interop_profile['UseCases']:
-            entry_title = use_case.get("UseCaseTitle", "NoName").replace(' ','_')
-            my_logger.debug('UseCase {}'.format(entry_title))
+            entry_title = use_case.get("UseCaseTitle", "NoName").replace(' ', '_')
+            entry_type = use_case.get("UseCaseType", "Normal")
+            my_logger.debug('UseCase {} {}'.format(entry_title, entry_type))
 
             # Check if we have a valid UseCase
-            if 'URIs' not in use_case and 'UseCaseKeyProperty' not in use_case:
+            if 'URIs' not in use_case and 'UseCaseKeyProperty' not in use_case and entry_type != 'AbsentResource':
                 my_logger.error('UseCase does not have URIs or UseCaseKeyProperty...')
 
-            if 'UseCaseKeyProperty' in use_case:
+            if entry_type == 'AbsentResource':
+                my_status = rf_payload.get('Status')
+                if my_status:
+                    use_case_applies = my_status.get('State') == 'Absent'
+                else:
+                    use_case_applies = False
+                if 'URIs' in use_case:
+                    use_case_applies = use_case_applies and checkInteropURI(propResourceObj, use_case['URIs'])
+
+            elif 'UseCaseKeyProperty' in use_case:
                 entry_key, entry_comparison, entry_values = use_case['UseCaseKeyProperty'], use_case['UseCaseComparison'], use_case['UseCaseKeyValues']
 
-                _, use_case_applies = checkComparison(rf_payload.get(entry_key), entry_comparison, entry_values)
+                _, use_case_applies = checkComparison(rf_payload.get(entry_key, REDFISH_ABSENT), entry_comparison, entry_values)
 
                 # Check if URI applies to this usecase as well
                 if 'URIs' in use_case:
@@ -702,12 +809,12 @@ def validateInteropResource(propResourceObj, interop_profile, rf_payload):
 
             elif 'URIs' in use_case:
                 use_case_applies = checkInteropURI(propResourceObj, use_case['URIs'])
-            
+
             else:
                 use_case_applies = False
 
             if use_case_applies:
-                my_msg = msgInterop("UseCase.{}".format(entry_title), '-', '-', '-', sEnum.OK)
+                my_msg = msgInterop("UseCase.{}".format(entry_title), '-', '-', '-', testResultEnum.OK)
 
                 msgs.append(my_msg)
 
@@ -718,8 +825,8 @@ def validateInteropResource(propResourceObj, interop_profile, rf_payload):
 
                 new_msgs, new_counts = validateInteropResource(propResourceObj, new_case, rf_payload)
 
-                if any([msg.success == sEnum.FAIL for msg in new_msgs]):
-                    my_msg.success = sEnum.FAIL
+                if any([msg.success == testResultEnum.FAIL for msg in new_msgs]):
+                    my_msg.success = testResultEnum.FAIL
 
                 msgs.extend(new_msgs)
                 counts.update(new_counts)
@@ -728,7 +835,6 @@ def validateInteropResource(propResourceObj, interop_profile, rf_payload):
                 my_logger.info('UseCase {} does not apply'.format(entry_title))
 
         return msgs, counts
-
     if "URIs" in interop_profile:
         # Check if the profile requirements apply to this particular instance
         if not checkInteropURI(propResourceObj, interop_profile['URIs']):
@@ -773,11 +879,11 @@ def validateInteropResource(propResourceObj, interop_profile, rf_payload):
         pass
 
     for item in [item for item in msgs if not item.ignore]:
-        if item.success == sEnum.WARN:
+        if item.success == testResultEnum.WARN:
             counts['warn'] += 1
-        elif item.success == sEnum.PASS:
+        elif item.success == testResultEnum.PASS:
             counts['pass'] += 1
-        elif item.success == sEnum.FAIL:
+        elif item.success == testResultEnum.FAIL:
             counts['fail.{}'.format(item.name)] += 1
         counts['totaltests'] += 1
     return msgs, counts